(feat) user name value object added

+++ src/main/java/com/edu/pe/gocareerapi/User/infrastructure/repositories/UserRepository.java
@@ -4,9 +4,7 @@ import com.edu.pe.gocareerapi.User.domain.model.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-
-
-}
+public interface UserRepository extends JpaRepository<User, Integer> {}
